refactor: 게임 메타데이터 관련 엔티티와 리포지토리를 GameData 도메인으로 분리
