--- src/main/java/com/summoner/lolhaeduo/client/repository/ProfileIconRepository.java
@@ -1,8 +1,0 @@
-package com.summoner.lolhaeduo.client.repository;
-
-import com.summoner.lolhaeduo.client.entity.Champion;
-import com.summoner.lolhaeduo.client.entity.ProfileIcon;
-import org.springframework.data.jpa.repository.JpaRepository;
-
-public interface ProfileIconRepository extends JpaRepository<ProfileIcon, Long> {
-}
