+++ src/main/java/com/summoner/lolhaeduo/client/service/RiotClientService.java
@@ -3,7 +3,7 @@ package com.summoner.lolhaeduo.client.service;
 import com.summoner.lolhaeduo.client.dto.*;
 import com.summoner.lolhaeduo.client.entity.Favorite;
 import com.summoner.lolhaeduo.client.repository.FavoriteRepository;
-import com.summoner.lolhaeduo.client.repository.VersionRepository;
+import com.summoner.lolhaeduo.domain.gamedata.repository.VersionRepository;
 import com.summoner.lolhaeduo.client.riot.RiotClient;
 import com.summoner.lolhaeduo.common.util.TimeUtil;
 import com.summoner.lolhaeduo.domain.account.dto.LinkAccountRequest;
