+++ src/main/java/com/summoner/lolhaeduo/domain/gamedata/repository/ProfileIconRepository.java
@@ -1,0 +1,7 @@
+package com.summoner.lolhaeduo.domain.gamedata.repository;
+
+import com.summoner.lolhaeduo.domain.gamedata.entity.ProfileIcon;
+import org.springframework.data.jpa.repository.JpaRepository;
+
+public interface ProfileIconRepository extends JpaRepository<ProfileIcon, Long> {
+}
