+++ src/main/java/com/summoner/lolhaeduo/domain/gamedata/repository/VersionRepository.java
@@ -1,6 +1,6 @@
-package com.summoner.lolhaeduo.client.repository;
+package com.summoner.lolhaeduo.domain.gamedata.repository;
 
-import com.summoner.lolhaeduo.client.entity.Version;
+import com.summoner.lolhaeduo.domain.gamedata.entity.Version;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.Query;
