+++ src/test/java/com/summoner/lolhaeduo/client/service/RiotClientServiceTest.java
@@ -1,7 +1,7 @@
 //package com.summoner.lolhaeduo.client.service;
 //
 //import com.summoner.lolhaeduo.client.repository.FavoriteRepository;
-//import com.summoner.lolhaeduo.client.repository.VersionRepository;
+//import com.summoner.lolhaeduo.domain.gamedata.repository.VersionRepository;
 //import com.summoner.lolhaeduo.client.riot.RiotClient;
 //import com.summoner.lolhaeduo.common.util.TimeUtil;
 //import com.summoner.lolhaeduo.domain.account.entity.Account;
